pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pysqueeze"
version = "0.2.0"
description = "High-quality image compression like TinyPNG and Squoosh"
authors = [{ name = "Gerard Nwazk", email = "gerardnwazk@hotmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "Pillow>=12.0.0",
    "pillow-avif-plugin>=1.4.0"
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
pysqueeze = "pysqueeze.cli:main"
